% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pebba.R
\name{get_cutoff}
\alias{get_cutoff}
\title{Get cutoff value}
\usage{
get_cutoff(deg_list, logFC_col, pvalue_col, top_n, min_genes, max_genes)
}
\arguments{
\item{deg_list}{A list of DEGs}

\item{logFC_col}{A string indicating the column with log fold
change values}

\item{pvalue_col}{A string indicating the column with p-values}

\item{top_n}{Top N something}

\item{min_genes}{Minimum number of genes}

\item{max_genes}{Maximum number of genes}
}
\description{
This function provides the cutoff value from FAIRcc
}
