% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pebba.R
\name{save_heatmap}
\alias{save_heatmap}
\title{Save heatmap}
\usage{
save_heatmap(df2heat, replace_p, p_cut, f_out, results_dir)
}
\arguments{
\item{df2heat}{Object to be turned to heatmap.}

\item{replace_p}{Replace P.}

\item{p_cut}{P-value cut.}

\item{f_out}{Output file name.}

\item{results_dir}{The path into which results should be saved (Default: "Results").}
}
\description{
This function is used to save heatmaps
}
\examples{
#Add examples here
print(save_heatmap)
}
